+++ src/main.cpp
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/01 17:15:45 by emadriga          #+#    #+#             */
-/*   Updated: 2023/04/11 20:35:09 by emadriga         ###   ########.fr       */
+/*   Updated: 2023/04/13 11:38:18 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,9 @@
 int main(int argc, char **argv)
 {
 	// atexit(ft_leaks);
+	if (argc > 1 && !std::strcmp(argv[1], "granular"))
 	{
+		ft::SelectServer T(AF_INET, SOCK_STREAM, 0, INADDR_ANY, 8080, 5);
 		std::cout << "Starting..." << std::endl;
 		// std::cout << "Socket..." << std::endl;
 		// ft::ASocket as = ft::ASocket(AF_INET, SOCK_STREAM, 0, INADDR_ANY, 80);